Refactor empty & sleepy inspections
