--- plugin/src/main/kotlin/org/jetbrains/research/pynose/plugin/inspections/pytest/EmptyTestTestSmellPytestInspection.kt
@@ -1,17 +1,0 @@
-package org.jetbrains.research.pynose.plugin.inspections.pytest
-
-import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.diagnostic.Logger
-import com.intellij.psi.PsiElementVisitor
-import org.jetbrains.research.pynose.plugin.inspections.AbstractTestSmellInspection
-import org.jetbrains.research.pynose.plugin.inspections.common.EmptyTestTestSmellInspectionVisitor
-
-
-class EmptyTestTestSmellPytestInspection : AbstractTestSmellInspection() {
-    private val LOG = Logger.getInstance(EmptyTestTestSmellPytestInspection::class.java)
-
-    override fun buildPytestVisitor(holder: ProblemsHolder, session: LocalInspectionToolSession): PsiElementVisitor {
-        return EmptyTestTestSmellInspectionVisitor(holder, session)
-    }
-}
