--- plugin/src/main/kotlin/org/jetbrains/research/pynose/plugin/inspections/unittest/EmptyTestTestSmellUnittestInspection.kt
@@ -1,17 +1,0 @@
-package org.jetbrains.research.pynose.plugin.inspections.unittest
-
-import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.diagnostic.Logger
-import com.intellij.psi.PsiElementVisitor
-import org.jetbrains.research.pynose.plugin.inspections.AbstractTestSmellInspection
-import org.jetbrains.research.pynose.plugin.inspections.common.EmptyTestTestSmellInspectionVisitor
-
-
-class EmptyTestTestSmellUnittestInspection : AbstractTestSmellInspection() {
-    private val LOG = Logger.getInstance(EmptyTestTestSmellUnittestInspection::class.java)
-
-    override fun buildUnittestVisitor(holder: ProblemsHolder, session: LocalInspectionToolSession): PsiElementVisitor {
-        return EmptyTestTestSmellInspectionVisitor(holder, session)
-    }
-}
