+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -87,17 +87,9 @@
                          groupName="PyNose"
                          enabledByDefault="true"
                          level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.unittest.EmptyTestTestSmellUnittestInspection"
-                         displayName="Empty test unittest"
-                         shortName="EmptyTestTestSmellUnittestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="true"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.pytest.EmptyTestTestSmellPytestInspection"
-                         displayName="Empty test pytest"
-                         shortName="EmptyTestTestSmellPytestInspection"/>
+                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.universal.EmptyTestTestSmellInspection"
+                         displayName="Empty test"
+                         shortName="EmptyTestTestSmellInspection"/>
         <localInspection language="Python"
                          groupPath="Python"
                          groupName="PyNose"
@@ -158,18 +150,10 @@
                          groupPath="Python"
                          groupName="PyNose"
                          enabledByDefault="false"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.unittest.disabled.SleepyTestTestSmellUnittestInspection"
-                         displayName="Sleepy test (unittest)"
-                         shortName="SleepyTestTestSmellUnittestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="false"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.pytest.disabled.SleepyTestTestSmellPytestInspection"
-                         displayName="Sleepy test (pytest)"
-                         shortName="SleepyTestTestSmellPytestInspection"/>
+                         level="WEAK WARNING"
+                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.universal.SleepyTestTestSmellInspection"
+                         displayName="Sleepy test"
+                         shortName="SleepyTestTestSmellInspection"/>
         <localInspection language="Python"
                          groupPath="Python"
                          groupName="PyNose"
