+++ plugin/src/main/resources/inspectionDescriptions/EmptyTestTestSmellInspection.html
