--- plugin/src/main/resources/inspectionDescriptions/EmptyTestTestSmellPytestInspection.html
@@ -1,8 +1,0 @@
-<html>
-<body>
-
-<b>Empty Test (pytest)</b> occurs when a test case does not contain executable statements.
-Such tests are possibly created for debugging purposes and then forgotten about or contain commented out code
-
-</body>
-</html>
