+++ plugin/src/main/resources/inspectionDescriptions/SleepyTestTestSmellInspection.html
