--- plugin/src/main/resources/inspectionDescriptions/SleepyTestTestSmellPytestInspection.html
@@ -1,10 +1,0 @@
-<html>
-<body>
-
-<b>Sleepy Test (pytest)</b> occurs when developers need to pause the execution of statements in a test case
-for a certain duration (i.e., simulate an external event) and then continue with the execution.
-Explicitly causing a thread to sleep can lead to unexpected results
-as the processing time for a task can vary on different devices.
-
-</body>
-</html>
