+++ plugin/src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/sleepy/pytest/test_sleepy_multiple.py
@@ -5,17 +5,17 @@ import time
 class TestClass(unittest.TestCase):
 
     def test_something(self):
-        <warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</warning>
+        <weak_warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</weak_warning>
 
     def do_something(self):
         time.sleep(5)
 
     def test_something_else(self):
-        <warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</warning>
+        <weak_warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</weak_warning>
 
 
 def test_something_other(self):
-    <warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</warning>
+    <weak_warning descr="Causing a thread to sleep can lead to different results on different devices, consider removing or documenting it">time.sleep(5)</weak_warning>
 
 
 class AnotherClass:
